+++ vite.config.ts
@@ -1,5 +1,6 @@
 import adapter from '@sveltejs/adapter-auto';
 import { sveltekit } from '@sveltejs/kit/vite';
+import tailwindcss from '@tailwindcss/vite';
 import { defineConfig } from 'vite';
 import glsl from 'vite-plugin-glsl';
 
@@ -19,6 +20,7 @@ export default defineConfig({
 		}
 	},
 	plugins: [
+		tailwindcss(),
 		glsl(),
 		sveltekit({
 			compilerOptions: {